delete separate exception files
