+++ arborio/xmlwrap.hpp
@@ -14,8 +14,6 @@
 #include <libxml/xpath.h>
 #include <libxml/xpathInternals.h>
 
-#include <arborio/nmlexcept.hpp>
-
 namespace arborio {
 
 // `non_negative` represents the corresponding constraint in the schema, which
